Use (preferred - minimum) instead of just preferred as weights.
The general idea of the proposed algorithm is as follows:

* If the sum of the preferred widths is small enough for
  the containing block, use that
* If the sum of the preferred minimum widths is too big,
  use that
* Otherwise, do something proportional to the amount of
  content.

For the proportional distribution, the algorithm used
to take the preferred widths as weights. The problem
is that this ignored the contraints of minimum widths.

In the new algorithm, each box always gets at least its
preferred mininmum width. In the third case, only the
remaining space is distributed. The weights are
the preferred width minus the preferred mininum width.

This new algorithm as better continuity while still
always keeping the mininum contsraints.